% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamplesSQLite3Proj.R
\name{getSamplesSQLite3Proj}
\alias{getSamplesSQLite3Proj}
\title{getSamplesSQLite3Proj}
\usage{
getSamplesSQLite3Proj(projID, dbPath, csvOut)
}
\arguments{
\item{projID}{project ID String for query, e.g. "Wolf analysis"}

\item{dbPath}{path to database, e.g. "D:/Data/db/anaLabDBtest/anaLabDB.db"}

\item{csvOut}{path for csv, e.g. "../dat/csv/samples.csv"}
}
\description{
Query the SQLite3 anaLabDB database for all samples
associated with a specific project.
}
\examples{
# not run
}
